feat: 세션 기반으로 Repository 변경
