--- src/main/java/matchuri/backend/global/security/HttpCookieOAuth2AuthorizationRequestRepository.java
@@ -1,66 +1,0 @@
-package matchuri.backend.global.security;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import matchuri.backend.global.config.MatchuriProperties;
-import org.springframework.security.oauth2.client.web.AuthorizationRequestRepository;
-import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class HttpCookieOAuth2AuthorizationRequestRepository
-        implements AuthorizationRequestRepository<OAuth2AuthorizationRequest> {
-
-    private final MatchuriProperties matchuriProperties;
-
-    @Override
-    public OAuth2AuthorizationRequest loadAuthorizationRequest(HttpServletRequest request) {
-        return CookieUtils.getCookie(request, cookieName())
-                .map(cookie -> CookieUtils.deserialize(cookie, OAuth2AuthorizationRequest.class))
-                .orElse(null);
-    }
-
-    @Override
-    public void saveAuthorizationRequest(
-            OAuth2AuthorizationRequest authorizationRequest,
-            HttpServletRequest request,
-            HttpServletResponse response
-    ) {
-        if (authorizationRequest == null) {
-            removeAuthorizationRequestCookies(request, response);
-            return;
-        }
-
-        CookieUtils.addCookie(
-                response,
-                cookieName(),
-                CookieUtils.serialize(authorizationRequest),
-                matchuriProperties.getAuth().getCookie().getOauth2AuthorizationRequestCookieMaxAgeSeconds(),
-                matchuriProperties.getAuth().getCookie()
-        );
-    }
-
-    @Override
-    public OAuth2AuthorizationRequest removeAuthorizationRequest(
-            HttpServletRequest request,
-            HttpServletResponse response
-    ) {
-        OAuth2AuthorizationRequest authorizationRequest = loadAuthorizationRequest(request);
-        removeAuthorizationRequestCookies(request, response);
-        return authorizationRequest;
-    }
-
-    public OAuth2AuthorizationRequest removeAuthorizationRequest(HttpServletRequest request) {
-        return loadAuthorizationRequest(request);
-    }
-
-    public void removeAuthorizationRequestCookies(HttpServletRequest request, HttpServletResponse response) {
-        CookieUtils.deleteCookie(response, cookieName(), matchuriProperties.getAuth().getCookie());
-    }
-
-    private String cookieName() {
-        return matchuriProperties.getAuth().getCookie().getOauth2AuthorizationRequestCookieName();
-    }
-}
