+++ src/main/java/matchuri/backend/global/security/MatchuriOAuth2AuthorizationRequestRepository.java
@@ -1,0 +1,42 @@
+package matchuri.backend.global.security;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.security.oauth2.client.web.AuthorizationRequestRepository;
+import org.springframework.security.oauth2.client.web.HttpSessionOAuth2AuthorizationRequestRepository;
+import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
+import org.springframework.stereotype.Component;
+
+@Component
+public class MatchuriOAuth2AuthorizationRequestRepository
+        implements AuthorizationRequestRepository<OAuth2AuthorizationRequest> {
+
+    private final HttpSessionOAuth2AuthorizationRequestRepository delegate =
+            new HttpSessionOAuth2AuthorizationRequestRepository();
+
+    @Override
+    public OAuth2AuthorizationRequest loadAuthorizationRequest(HttpServletRequest request) {
+        return delegate.loadAuthorizationRequest(request);
+    }
+
+    @Override
+    public void saveAuthorizationRequest(
+            OAuth2AuthorizationRequest authorizationRequest,
+            HttpServletRequest request,
+            HttpServletResponse response
+    ) {
+        delegate.saveAuthorizationRequest(authorizationRequest, request, response);
+    }
+
+    @Override
+    public OAuth2AuthorizationRequest removeAuthorizationRequest(
+            HttpServletRequest request,
+            HttpServletResponse response
+    ) {
+        return delegate.removeAuthorizationRequest(request, response);
+    }
+
+    public void clearAuthorizationRequest(HttpServletRequest request, HttpServletResponse response) {
+        delegate.removeAuthorizationRequest(request, response);
+    }
+}
